♻️ Refactor: 폴더명 컨벤션 따라서 수정 (main step)
