+++ src/common/components/box-text/SnsLink/SnsLink.client.tsx
