+++ src/common/components/box-text/SnsLink/SnsLink.types.ts
