+++ src/common/components/box-text/SurveyInfo/SurveyInfo.server.tsx
