+++ src/common/components/box-text/SurveyInfo/SurveyInfo.types.ts
