+++ src/common/components/list-file/ListFileSnsEventAssistant/ListFileSnsEventAssistant.client.tsx
