+++ src/common/components/list-file/ListFileSnsEventAssistant/ListFileSnsEventAssistant.types.ts
