+++ src/common/components/list-file/ListFileSnsEventAssistantLink/ListFileSnsEventAssistantLink.types.ts
