+++ src/common/components/list-file/ListFileTeamMoodTracker/ListFileTeamMoodTracker.stories.tsx
