+++ src/common/components/survey/BarChart/BarChart.stories.tsx
