+++ src/common/components/survey/BarChart/BarChart.types.ts
