+++ src/common/components/survey/BarChart/bar-chart-config.ts
