+++ src/common/components/survey/BarChart/bar-chart-plugin.ts
