+++ src/common/components/survey/PieChart/pie-chart-config.ts
