+++ src/common/components/survey/PieChart/pie-chart-plugin.ts
