+++ src/common/components/survey/SubjectiveAnswer/SubjectiveAnswser.server.tsx
@@ -1,4 +1,4 @@
-import ChartContainer from '@common/components/survey-temp/ChartContainer/ChartContainer.server';
+import ChartContainer from '@common/components/survey/ChartContainer/ChartContainer.server';
 
 import { SubjectiveAnswersProps } from './SubjectiveAnswer.types';
